split feast package into subpackages
