+++ go/internal/feast/onlineserving/serving_test.go
@@ -1,0 +1,180 @@
+package onlineserving
+
+import (
+	"github.com/feast-dev/feast/go/internal/feast/model"
+	"github.com/feast-dev/feast/go/protos/feast/types"
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+func TestGroupingFeatureRefs(t *testing.T) {
+	viewA := &model.FeatureView{
+		Base: &model.BaseFeatureView{
+			Name: "viewA",
+			Projection: &model.FeatureViewProjection{
+				NameAlias: "aliasViewA",
+			},
+		},
+		Entities: map[string]struct{}{"driver": {}, "customer": {}},
+	}
+	viewB := &model.FeatureView{
+		Base:     &model.BaseFeatureView{Name: "viewB"},
+		Entities: map[string]struct{}{"driver": {}, "customer": {}},
+	}
+	viewC := &model.FeatureView{
+		Base:     &model.BaseFeatureView{Name: "viewC"},
+		Entities: map[string]struct{}{"driver": {}},
+	}
+	viewD := &model.FeatureView{
+		Base:     &model.BaseFeatureView{Name: "viewD"},
+		Entities: map[string]struct{}{"customer": {}},
+	}
+	refGroups, _ := GroupFeatureRefs(
+		[]*FeatureViewAndRefs{
+			{View: viewA, FeatureRefs: []string{"featureA", "featureB"}},
+			{View: viewB, FeatureRefs: []string{"featureC", "featureD"}},
+			{View: viewC, FeatureRefs: []string{"featureE"}},
+			{View: viewD, FeatureRefs: []string{"featureF"}},
+		},
+		map[string]*types.RepeatedValue{
+			"driver_id": {Val: []*types.Value{
+				{Val: &types.Value_Int32Val{Int32Val: 0}},
+				{Val: &types.Value_Int32Val{Int32Val: 0}},
+				{Val: &types.Value_Int32Val{Int32Val: 1}},
+				{Val: &types.Value_Int32Val{Int32Val: 1}},
+				{Val: &types.Value_Int32Val{Int32Val: 1}},
+			}},
+			"customer_id": {Val: []*types.Value{
+				{Val: &types.Value_Int32Val{Int32Val: 1}},
+				{Val: &types.Value_Int32Val{Int32Val: 2}},
+				{Val: &types.Value_Int32Val{Int32Val: 3}},
+				{Val: &types.Value_Int32Val{Int32Val: 3}},
+				{Val: &types.Value_Int32Val{Int32Val: 4}},
+			}},
+		},
+		map[string]string{
+			"driver":   "driver_id",
+			"customer": "customer_id",
+		},
+		true,
+	)
+
+	assert.Len(t, refGroups, 3)
+
+	// Group 1
+	assert.Equal(t, []string{"featureA", "featureB", "featureC", "featureD"},
+		refGroups["customer_id,driver_id"].FeatureNames)
+	assert.Equal(t, []string{"viewA", "viewA", "viewB", "viewB"},
+		refGroups["customer_id,driver_id"].FeatureViewNames)
+	assert.Equal(t, []string{
+		"aliasViewA__featureA", "aliasViewA__featureB",
+		"viewB__featureC", "viewB__featureD"},
+		refGroups["customer_id,driver_id"].AliasedFeatureNames)
+	for _, group := range [][]int{{0}, {1}, {2, 3}, {4}} {
+		assert.Contains(t, refGroups["customer_id,driver_id"].Indices, group)
+	}
+
+	// Group2
+	assert.Equal(t, []string{"featureE"},
+		refGroups["driver_id"].FeatureNames)
+	for _, group := range [][]int{{0, 1}, {2, 3, 4}} {
+		assert.Contains(t, refGroups["driver_id"].Indices, group)
+	}
+
+	// Group3
+	assert.Equal(t, []string{"featureF"},
+		refGroups["customer_id"].FeatureNames)
+
+	for _, group := range [][]int{{0}, {1}, {2, 3}, {4}} {
+		assert.Contains(t, refGroups["customer_id"].Indices, group)
+	}
+
+}
+
+func TestGroupingFeatureRefsWithJoinKeyAliases(t *testing.T) {
+	viewA := &model.FeatureView{
+		Base: &model.BaseFeatureView{
+			Name: "viewA",
+			Projection: &model.FeatureViewProjection{
+				Name:       "viewA",
+				JoinKeyMap: map[string]string{"location_id": "destination_id"},
+			},
+		},
+		Entities: map[string]struct{}{"location": {}},
+	}
+	viewB := &model.FeatureView{
+		Base:     &model.BaseFeatureView{Name: "viewB"},
+		Entities: map[string]struct{}{"location": {}},
+	}
+
+	refGroups, _ := GroupFeatureRefs(
+		[]*FeatureViewAndRefs{
+			{View: viewA, FeatureRefs: []string{"featureA", "featureB"}},
+			{View: viewB, FeatureRefs: []string{"featureC", "featureD"}},
+		},
+		map[string]*types.RepeatedValue{
+			"location_id": {Val: []*types.Value{
+				{Val: &types.Value_Int32Val{Int32Val: 0}},
+				{Val: &types.Value_Int32Val{Int32Val: 0}},
+				{Val: &types.Value_Int32Val{Int32Val: 1}},
+				{Val: &types.Value_Int32Val{Int32Val: 1}},
+				{Val: &types.Value_Int32Val{Int32Val: 1}},
+			}},
+			"destination_id": {Val: []*types.Value{
+				{Val: &types.Value_Int32Val{Int32Val: 1}},
+				{Val: &types.Value_Int32Val{Int32Val: 2}},
+				{Val: &types.Value_Int32Val{Int32Val: 3}},
+				{Val: &types.Value_Int32Val{Int32Val: 3}},
+				{Val: &types.Value_Int32Val{Int32Val: 4}},
+			}},
+		},
+		map[string]string{
+			"location": "location_id",
+		},
+		true,
+	)
+
+	assert.Len(t, refGroups, 2)
+
+	assert.Equal(t, []string{"featureA", "featureB"},
+		refGroups["location_id[destination_id]"].FeatureNames)
+	for _, group := range [][]int{{0}, {1}, {2, 3}, {4}} {
+		assert.Contains(t, refGroups["location_id[destination_id]"].Indices, group)
+	}
+
+	assert.Equal(t, []string{"featureC", "featureD"},
+		refGroups["location_id"].FeatureNames)
+	for _, group := range [][]int{{0, 1}, {2, 3, 4}} {
+		assert.Contains(t, refGroups["location_id"].Indices, group)
+	}
+
+}
+
+func TestGroupingFeatureRefsWithMissingKey(t *testing.T) {
+	viewA := &model.FeatureView{
+		Base: &model.BaseFeatureView{
+			Name: "viewA",
+			Projection: &model.FeatureViewProjection{
+				Name:       "viewA",
+				JoinKeyMap: map[string]string{"location_id": "destination_id"},
+			},
+		},
+		Entities: map[string]struct{}{"location": {}},
+	}
+
+	_, err := GroupFeatureRefs(
+		[]*FeatureViewAndRefs{
+			{View: viewA, FeatureRefs: []string{"featureA", "featureB"}},
+		},
+		map[string]*types.RepeatedValue{
+			"location_id": {Val: []*types.Value{
+				{Val: &types.Value_Int32Val{Int32Val: 0}},
+			}},
+		},
+		map[string]string{
+			"location": "location_id",
+		},
+		true,
+	)
+	assert.Errorf(t, err, "key destination_id is missing in provided entity rows")
+}
