+++ go/internal/feast/registry/local.go
@@ -1,4 +1,4 @@
-package feast
+package registry
 
 import (
 	"github.com/feast-dev/feast/go/protos/feast/core"
